Deployed from fix: update the design to accommodate improved miss diagnostics.
This commit updates the design of call caching such that only the "identify" of
the task is considered for the cache key.

The command, requirements, hints, backend inputs, etc. are stored with digests
in the cache entry of the file.

This enables Sprocket to better report on _why_ it is ignoring a call cache
hit, such as the command changed or a particular input changed.

The previous design prevented this by basing the cache key upon the hash of
_everything_ relating to the task's execution. We're trading the cache entry
being sensitive to the location and name of the task in for the ability to
specifically report on why a cache entry is no longer valid. 99ab788a8fb21b04ebac8fb63977b2ea91ebe6f2